+++ src/lib/error.js
@@ -1,4 +1,5 @@
 // erros do firebase //
+
 export const errorsFirebase = (errorCode) => {
   switch (errorCode) {
     case 'auth/invalid-display-name':
@@ -18,41 +19,9 @@ export const errorsFirebase = (errorCode) => {
   }
 };
 
-// validar nome do perfil, email e senha do cadastro//
-export const validateFormRegister = (nameProfile, email, password) => {
-  // const passwordRegEx = /^(?=.*[a-z])(?=.*[A-Z])[0-9a-zA-Z]{6}/;
-  if (nameProfile === '') {
-    return 'Digite seu nome de perfil.';
+export const errorsFirebaseFirestore = (errorCode) => {
+  switch (errorCode) {
+    default:
+      return `Algo inexperado aconteceu. Por favor entre em contato com o suporte e informe o código ${errorCode}`;
   }
-  if (email === '') {
-    return 'Digite seu email';
-  }
-  if (password === '') {
-    return 'Digite sua senha';
-  }
-  // if (password !== passwordRegEx) {
-  //   return 'Sua senha deve ser de no minimo 6 digitos com letras e números.';
-  //   return messageError;
-  // }
-  return '';
-};
-
-// validar email e senha do login//
-export const validateFormlogin = (email, password) => {
-  let messageError;
-  // const emailRegEx = /^(?=.*[a-z])(?=.*[A-Z])[0-9a-zA-Z]{6}/;
-
-  if (email === '') {
-    messageError = 'Digite seu email';
-    return messageError;
-  }
-  if (password === '') {
-    messageError = 'Digite sua senha';
-    return messageError;
-  }
-  // if (password !== emailRegEx) {
-  //   messageError = 'Sua senha deve ser de no minimo 6 digitos com letras e números.';
-  //   return messageError;
-  // }
-  return '';
 };
