+++ src/pages/register/register.js
@@ -1,5 +1,6 @@
 import { register } from '../../lib/auth.js';
-import { errorsFirebase, validateFormRegister } from '../../lib/error.js';
+import { errorsFirebase } from '../../lib/error.js';
+import { validateFormRegister } from '../../lib/authenticate.js';
 
 export default () => {
   const containerRegistration = document.createElement('div');
@@ -13,14 +14,13 @@ export default () => {
       </figure>
 
       <forms class="registration-input">
-        <p id= 'message-welcome' class = 'walcome-message' role='dialog'> </p>
+        <p id= 'error-message' class = 'error-message'> </p>
+        <p id= 'message-welcome' class = 'walcome-message'> </p>
         <input type="text" name="profile-name" id="input-profile-name" class="input" placeholder="Nome do perfil " required>
         <input type="email" name="email" id="input-email-registration" class="input" placeholder="Digite seu email". required>
         <input type="password" id="password" class="input" placeholder="Senha de 6 dígitos" required>
       </forms>
       <button type="button" id="button-registration" class="button">Finalizar cadastro</button>
-
-      <p id= 'error-message' class = 'error-message'> </p>
   
     </main>
   
@@ -41,10 +41,12 @@ export default () => {
 
   btnRegistration.addEventListener('click', (e) => {
     e.preventDefault();
-    messageError.innerHTML = '';
     const validate = validateFormRegister(nameProfile.value, email.value, password.value);
     if (validate) {
       messageError.innerHTML = validate;
+      setTimeout(() => {
+        messageError.innerHTML = '';
+      }, 2000);
     } else {
       register(email.value, password.value, nameProfile.value)
         .then(() => {
@@ -56,6 +58,9 @@ export default () => {
         .catch((error) => {
           const errorCode = errorsFirebase(error.code);
           messageError.innerHTML = errorCode;
+          setTimeout(() => {
+            messageError.innerHTML = '';
+          }, 2000);
         });
     }
   });
